tests/requirements-test.txt
Django==5.0
django-pwa==1.1.0
fontawesomefree==6.4.2
gunicorn==21.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pytest==8.3.4
pytest-django==4.9.0
apache-airflow==2.10.3
Jinja2>=3.1.2
requests
sqlalchemy
apache-airflow-providers-postgres>=5.10.0
streamlit
pandas==2.2.3
polars==1.17.1
